Version 2.0.0
Created variables in pom.xml for easier maintenance.

Created a InputListener that listens for when the user presses enter
with a corresponding interface (IStoppableThread).

The PathOfExile package has been removed and reworked into the class
and interface IMessageHandler/MessageHandler.
The interface is used as a callback from IMAPFolder for when a message
is found.
MessageHandler prints and copies the code if it is found.

IMAPFolder instead of returning an array of Message, sends them to
the MessageHandler class via callback.

If the latest message form GGG is sent for more than 3 minutes ago,
the program will start a thread that idles until an new message arrives.

Most of the logic in main is moved to the above classes as a result.